Move all utility functions to utils.js
